+++ lightning_pose/models/heads/regression.py
@@ -1,5 +1,6 @@
 """Heads that produce (x, y) predictions for coordinate regression."""
 
+from lightning.pytorch import LightningModule
 from torch import nn
 from torchtyping import TensorType
 
@@ -9,7 +10,7 @@ __all__ = [
 ]
 
 
-class LinearRegressionHead(nn.Module):
+class LinearRegressionHead(LightningModule):
     """Linear regression head that converts 2D feature maps to a vector of (x, y) coordinates."""
 
     def __init__(
