+++ lightning_pose/utils/scripts.py
@@ -386,7 +386,6 @@ def get_model(
                 backbone=cfg.model.backbone,
                 pretrained=backbone_pretrained,
                 downsample_factor=cfg.data.get("downsample_factor", 2),
-                output_shape=data_module.dataset.output_shape,
                 torch_seed=cfg.training.rng_seed_model_pt,
                 optimizer=optimizer,
                 optimizer_params=optimizer_params,
@@ -401,7 +400,6 @@ def get_model(
                 backbone=cfg.model.backbone,
                 pretrained=backbone_pretrained,
                 downsample_factor=cfg.data.get("downsample_factor", 2),
-                output_shape=data_module.dataset.output_shape,
                 torch_seed=cfg.training.rng_seed_model_pt,
                 optimizer=optimizer,
                 optimizer_params=optimizer_params,
@@ -439,7 +437,6 @@ def get_model(
                 backbone=cfg.model.backbone,
                 pretrained=backbone_pretrained,
                 downsample_factor=cfg.data.get("downsample_factor", 2),
-                output_shape=data_module.dataset.output_shape,
                 torch_seed=cfg.training.rng_seed_model_pt,
                 optimizer=optimizer,
                 optimizer_params=optimizer_params,
@@ -455,7 +452,6 @@ def get_model(
                 backbone=cfg.model.backbone,
                 pretrained=backbone_pretrained,
                 downsample_factor=cfg.data.get("downsample_factor", 2),
-                output_shape=data_module.dataset.output_shape,
                 torch_seed=cfg.training.rng_seed_model_pt,
                 optimizer=optimizer,
                 optimizer_params=optimizer_params,
