+++ tests/models/test_heatmap_tracker.py
